refactor(search): remove `SearchRepository` interface and update package references
